decode: fix transparent background bug
This PR fixes qoi-rust#16 where the instructions [RUN 1, RGB r g b,
INDEX 53] produced [Black, rgb, Transparent] instead of [Black, rgb,
Black]. 53 is the hash of (0, 0, 0, 255) and in the special case where
the first instruction is RUN it needs to store the color in the index so
that future INDEX instructions can fetch the value. This is only needed
if the first instruction is RUN.

Since this is only needed for the first instruction one could probably
handle this edgecase before the loop. If that is something we want, we
can either modify this PR or you can make a new one with a better
solution.

[ fix rebase conflict, update commit message ]
Signed-off-by: Marc-André Lureau <marcandre.lureau@redhat.com>